"add change credentials page, rewrite profile class and table, add page new Shipment and class,controler,repo,service, work function add new shipment"

+++ src/main/java/ua/moisak/PostService/repositories/ProfileRepository.java
@@ -3,6 +3,6 @@ package ua.moisak.PostService.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.moisak.PostService.models.Profile;
 
-public interface ProfileRepository extends JpaRepository<Profile, Long> {
+public interface ProfileRepository extends JpaRepository<Profile, Integer> {
     void deleteByEmail(String email);
 }
